msm: mdss: Clear PP pipe resources before copying new userspace config
During the pipe setup function, we copy any PP related parameters from
userspace to the pipe cache. We currently do not clear the cache
beforehand. In cases where the pipe is being reconfigured, stale, invalid
values might still be present in the cache since the last time the pipe
was configured. These values need to be cleared before accepting a new
config.

Change-Id: I0d9a51052bb738f599f386d39cb9de5509aae7ac
Signed-off-by: Benet Clark <benetc@codeaurora.org>
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>