gpio_keys: force sync on KEY_POWER, KEY_HOME press This should help in situations when the device or the OS is stalled but the kernel is alive and running.
Typically, user will press the power or the home button to see if the device is responding
or to do a force restart.

By forcing a sync on KEY_POWER or KEY_HOME pressed, it will flush all the caches and prevent data corruption
and further more, prevent boot loop in these situations.

All the rest situations, when the device is NOT stalled and fully operational,
it will still not be a bad thing since dyn_fsync is calling sys_sync() on suspend as well.

Signed-off-by: placiano <placiano80@gmail.com>